+++ packages/fluentui/docs/src/examples/components/Tree/Usage/TreeMultiselectExample.shorthand.steps.ts
@@ -2,6 +2,7 @@ import { treeItemClassName, treeTitleClassName, treeTitleSlotClassNames } from '
 
 const selectors = {
   treeTitle: (itemIndex: number) => `.${treeItemClassName}:nth-of-type(${itemIndex}) .${treeTitleClassName}`,
+  treeItem: (itemIndex: number) => `.${treeItemClassName}:nth-of-type(${itemIndex})`,
   selectionIndicator: (itemIndex: number) =>
     `.${treeItemClassName}:nth-of-type(${itemIndex}) .${treeTitleSlotClassNames.indicator}`,
 };
@@ -21,11 +22,19 @@ const config: ScreenerTestsConfig = {
         .click(selectors.selectionIndicator(12))
         .snapshot('selected, when clicked on selection indicator')
         .click(selectors.treeTitle(13))
-        .snapshot('selected, when clicked on tree title')
+        .snapshot('selected, when clicked on selection indicator')
+        .click(selectors.treeTitle(4))
+        .snapshot('selected, when group partially selected')
         .click(selectors.selectionIndicator(2))
         .snapshot('all children selected')
         .keys(selectors.treeTitle(7), keys.space)
-        .snapshot('selected, when space pressed'),
+        .snapshot('selected, when space pressed')
+        .click(selectors.treeTitle(15))
+        .keys(selectors.treeTitle(16), keys.space)
+        .keys(selectors.treeTitle(17), keys.space)
+        .snapshot('selected, when group has non selectable item')
+        .keys(selectors.treeItem(15), keys.space)
+        .snapshot('toggle group selected'),
   ],
 };