+++ packages/fluentui/react-northstar/src/components/Table/TableRow.tsx
@@ -1,29 +1,29 @@
-import { ReactAccessibilityBehavior } from '@fluentui/react-bindings';
-import * as customPropTypes from '@fluentui/react-proptypes';
+import { Accessibility, tableRowBehavior, GridRowBehaviorProps } from '@fluentui/accessibility';
+import { getElementType, useAccessibility, useStyles, useTelemetry, useUnhandledProps } from '@fluentui/react-bindings';
 import { Ref } from '@fluentui/react-component-ref';
-import * as PropTypes from 'prop-types';
+import * as customPropTypes from '@fluentui/react-proptypes';
+import { mergeComponentVariables } from '@fluentui/styles';
 import * as _ from 'lodash';
+// @ts-ignore
+import { ThemeContext } from 'react-fela';
+import * as PropTypes from 'prop-types';
 import * as React from 'react';
-import TableCell, { TableCellProps } from './TableCell';
 import {
-  UIComponent,
-  RenderResultConfig,
-  UIComponentProps,
-  commonPropTypes,
-  ShorthandFactory,
-  createShorthandFactory,
-  applyAccessibilityKeyHandlers,
-  childrenExist,
-} from '../../utils';
-import { ShorthandCollection, WithAsProp, withSafeTypeForAs } from '../../types';
-import { Accessibility, tableRowBehavior } from '@fluentui/accessibility';
-import { ComponentVariablesObject, mergeComponentVariables } from '@fluentui/styles';
+  FluentComponentStaticProps,
+  ProviderContextPrepared,
+  ShorthandCollection,
+  WithAsProp,
+  withSafeTypeForAs,
+  ComponentEventHandler,
+} from '../../types';
+import { childrenExist, commonPropTypes, createShorthandFactory, UIComponentProps } from '../../utils';
+import TableCell, { TableCellProps } from './TableCell';
 
 export interface TableRowProps extends UIComponentProps {
   /**
    * Accessibility behavior if overridden by the user.
    * */
-  accessibility?: Accessibility;
+  accessibility?: Accessibility<GridRowBehaviorProps>;
 
   /**
    * Row cells
@@ -44,6 +44,14 @@ export interface TableRowProps extends UIComponentProps {
    * Whether a row is currently selected or not.
    */
   selected?: boolean;
+
+  /**
+   * Called on click.
+   *
+   * @param event - React's original SyntheticEvent.
+   * @param data - All props.
+   */
+  onClick?: ComponentEventHandler<TableRowProps>;
 }
 
 const handleVariablesOverrides = variables => predefinedProps => ({
@@ -52,94 +60,108 @@ const handleVariablesOverrides = variables => predefinedProps => ({
 
 export const tableRowClassName = 'ui-table__row';
 
-class TableRow extends UIComponent<WithAsProp<TableRowProps>> {
-  static displayName = 'TableRow';
+export type TableRowStylesProps = Pick<TableRowProps, 'header' | 'compact'>;
 
-  static deprecated_className = tableRowClassName;
+const TableRow: React.FC<WithAsProp<TableRowProps>> & FluentComponentStaticProps<TableRowProps> = props => {
+  const context: ProviderContextPrepared = React.useContext(ThemeContext);
+  const { setStart, setEnd } = useTelemetry(TableRow.displayName, context.telemetry);
+  setStart();
+  const rowRef = React.useRef<HTMLElement>();
+  const { className, design, styles, items, header, compact, children, accessibility, variables, selected } = props;
 
-  static create: ShorthandFactory<TableRowProps>;
-
-  static propTypes = {
-    ...commonPropTypes.createCommon({
-      content: false,
+  const hasChildren = childrenExist(children);
+  const ElementType = getElementType(props);
+  const unhandledProps = useUnhandledProps(TableRow.handledProps, props);
+  const getA11yProps = useAccessibility(accessibility, {
+    debugName: TableRow.displayName,
+    actionHandlers: {
+      // https://github.com/microsoft/fluent-ui-react/issues/2150
+      unsetRowTabbable: e => {
+        rowRef.current.setAttribute('tabindex', '-1');
+      },
+      performClick: e => {
+        handleClick(e);
+      },
+    },
+    mapPropsToBehavior: () => ({
+      selected,
+      header,
     }),
-    content: customPropTypes.every([
-      customPropTypes.disallow(['children']),
-      PropTypes.oneOfType([PropTypes.arrayOf(customPropTypes.nodeContent), customPropTypes.nodeContent]),
-    ]),
-    items: customPropTypes.collectionShorthand,
-    header: PropTypes.bool,
-    compact: PropTypes.bool,
-    selected: PropTypes.bool,
-  };
+    rtl: context.rtl,
+  });
 
-  static defaultProps = {
-    accessibility: tableRowBehavior as Accessibility,
-  };
+  const { classes } = useStyles<TableRowStylesProps>(TableRow.displayName, {
+    className: tableRowClassName,
+    mapPropsToStyles: () => ({
+      header,
+      compact,
+    }),
+    mapPropsToInlineStyles: () => ({
+      className,
+      design,
+      styles,
+      variables,
+    }),
+    rtl: context.rtl,
+  });
 
-  rowRef = React.createRef<HTMLElement>();
-
-  actionHandlers = {
-    // https://github.com/microsoft/fluent-ui-react/issues/2150
-    unsetRowTabbable: e => {
-      this.rowRef.current.setAttribute('tabindex', '-1');
-    },
-    performClick: e => {
-      this.handleClick(e);
-    },
-  };
-
-  handleClick = (e: React.SyntheticEvent) => {
+  const handleClick = (e: React.SyntheticEvent) => {
     if (e.currentTarget === e.target) {
-      _.invoke(this.props, 'onClick', e, this.props);
+      _.invoke(props, 'onClick', e, props);
       e.preventDefault();
     }
   };
 
-  renderCells(accessibility: ReactAccessibilityBehavior, variables: ComponentVariablesObject) {
-    const { items } = this.props;
-
-    const cellAccessibility = accessibility.childBehaviors ? accessibility.childBehaviors.cell : undefined;
-
+  const renderCells = () => {
     return _.map(items, (item: TableCellProps, index: number) => {
-      const cellProps = {
-        accessibility: cellAccessibility as Accessibility,
-      };
       const overrideProps = handleVariablesOverrides(variables);
-
       return TableCell.create(item, {
-        defaultProps: () => cellProps,
-        overrideProps,
+        defaultProps: () =>
+          getA11yProps('cell', {
+            ...overrideProps,
+          }),
       });
     });
-  }
+  };
 
-  renderComponent({
-    accessibility,
-    ElementType,
-    classes,
-    variables,
-    unhandledProps,
-  }: RenderResultConfig<any>): React.ReactNode {
-    const { children } = this.props;
-    const hasChildren = childrenExist(children);
-
-    return (
-      <Ref innerRef={this.rowRef}>
+  const element = (
+    <Ref innerRef={rowRef}>
+      {getA11yProps.unstable_wrapWithFocusZone(
         <ElementType
-          className={classes.root}
-          onClick={this.handleClick}
-          {...accessibility.attributes.root}
-          {...unhandledProps}
-          {...applyAccessibilityKeyHandlers(accessibility.keyHandlers.root, unhandledProps)}
+          {...getA11yProps('root', {
+            className: classes.root,
+            onClick: handleClick,
+            ...unhandledProps,
+          })}
         >
           {hasChildren && children}
-          {!hasChildren && this.renderCells(accessibility, variables)}
-        </ElementType>
-      </Ref>
-    );
-  }
-}
+          {!hasChildren && renderCells()}
+        </ElementType>,
+      )}
+    </Ref>
+  );
+  setEnd();
+  return element;
+};
+
+TableRow.displayName = 'TableRow';
+
+TableRow.propTypes = {
+  ...commonPropTypes.createCommon({
+    content: false,
+  }),
+  items: customPropTypes.collectionShorthand,
+  header: PropTypes.bool,
+  compact: PropTypes.bool,
+  selected: PropTypes.bool,
+  onClick: PropTypes.func,
+};
+
+TableRow.handledProps = Object.keys(TableRow.propTypes) as any;
+
+TableRow.defaultProps = {
+  accessibility: tableRowBehavior,
+};
 
 TableRow.create = createShorthandFactory({ Component: TableRow, mappedArrayProp: 'items' });