+++ packages/fluentui/react-northstar/test/specs/components/TextArea/TextArea-test.tsx
@@ -5,9 +5,7 @@ import { isConformant } from 'test/specs/commonTests';
 import * as faker from 'faker';
 
 describe('TextArea', () => {
-  isConformant(TextArea, {
-    autoControlledProps: ['value'],
-  });
+  isConformant(TextArea, { constructorName: 'TextArea', autoControlledProps: ['value'] });
 
   describe('defaultValue', () => {
     test('sets "defaultValue" as initial "value"', () => {